[Feat] #6 - 타이머뷰 기능 및 UI 구현 & 인증코드뷰 및 타이머뷰 분리
