+++ playkuround-iOS/Views/Login/LoginView.swift
@@ -14,12 +14,7 @@ struct LoginView: View {
     @State private var mailButtonName = StringLiterals.Login.requestCode
     @State private var mailButtonClicked = false
     
-    @State private var certificationCode: String = ""
-    @State private var certificationButtonClicked = false
-    
     @State private var isShowingBottomSheet = false
-    
-    private let mailSystemURL = URL(string: StringLiterals.Login.mailSystemURL)
     
     var body: some View {
         ZStack {
@@ -74,47 +69,7 @@ struct LoginView: View {
                     }
                 
                 if mailButtonClicked {
-                    VStack(alignment: .leading) {
-                        Text(StringLiterals.Login.goEmail)
-                            .font(.pretendard12R)
-                            .foregroundStyle(.kuDarkBlue).underline()
-                            .padding(.top, 6)
-                            .onTapGesture {
-                                if let mailSystemURL = mailSystemURL {
-                                    UIApplication.shared.open(mailSystemURL)
-                                }
-                                else {
-                                    print("url error")
-                                }
-                            }
-                        
-                        Image(.longButtonWhite)
-                            .overlay {
-                                TextField(StringLiterals.Login.authenticationCode, text: $certificationCode)
-                                    .font(.pretendard15R)
-                                    .kerning(-0.41)
-                                    .padding(.leading, 20)
-                            }
-                            .padding(.top, 46)
-                        
-                        Image(certificationCode.isEmpty ? "longButtonGray" : "longButtonBlue")
-                            .onTapGesture {
-                                certificationButtonClicked.toggle()
-                                
-                                if certificationButtonClicked {
-                                    
-                                }
-                                else {
-                                    
-                                }
-                            }
-                            .overlay {
-                                Text(StringLiterals.Login.authentication)
-                                    .font(.neo15)
-                                    .foregroundStyle(.kuText)
-                                    .kerning(-0.41)
-                            }
-                    }
+                    AuthenticationCodeView()
                 }
                 
                 Spacer()
